Rename snippet files again
Renaming them to *.txt was not such a bright idea. The problem is that
some editors enable hard line-wrapping by default for .txt files, which
is not a good idea for source text of any kind.

Therefore rename the files again.

This time omit the file extension entirely, because I could not find a
good one so far. (I thought of *.snippet, but that's a lot to type in
unless tab completion is available.)